option for only positive distance reward for CarlaEnvironment
The agent still receives positive reward for moving ahead along the
route, but will not receive negative reward
for moving back along the route. Instead, the negative distance
will be accumulated to the future distance reward. This may ease
the learning if the right behavior is to temporarily go back along
the route in order, for examle, to avoid obstacle.